Fix summary duplication
