+++ janus_core/cli/geomopt.py
@@ -29,7 +29,6 @@ from janus_core.cli.utils import (
     start_summary,
     yaml_converter_callback,
 )
-from janus_core.helpers.utils import dict_paths_to_strs
 
 app = Typer()
 
@@ -263,11 +262,7 @@ def geomopt(
     # Store inputs for yaml summary
     inputs = optimize_kwargs.copy()
 
-    # Store only filename as filemode is not set by user
-    del inputs["log_kwargs"]
-    inputs["log"] = log
-
-    # Add structure and MLIP information to inputs
+    # Add structure, MLIP information, and log to inputs
     save_struct_calc(
         inputs=inputs,
         struct=optimizer.struct,
@@ -277,10 +272,8 @@ def geomopt(
         model_path=model_path,
         read_kwargs=read_kwargs,
         calc_kwargs=calc_kwargs,
+        log=log,
     )
-
-    # Convert all paths to strings in inputs nested dictionary
-    dict_paths_to_strs(inputs)
 
     # Save summary information before optimization begins
     start_summary(command="geomopt", summary=summary, inputs=inputs)
